These Matlab files help visualize the snake robot
drawlink is used to generate a link element of the robot.
drawstate, takes the state of the robot as input and plots all the links
of the robot by repeatedly calling drawlink function.
viz_try_1 function is a simple script that randomly simulates the motion
of the snake robot and visualizes it.
We can add cad model of the tube, organ etc in the plot.